+++ Changelog.md
@@ -7,6 +7,7 @@
 * Add Volume Conversions.
 * In the Quadratic Equation Solver, pressing Enter in one of the input fields has the same effect as clicking the Solve button.
 * Update the instructions in the Quadratic Equation Solver.
+* Update the interface of the Pythagorean Triple Finder.
 * Update the interface of the Quadratic Equation Solver.
 * Update the interface of the Rectangles window.
 * Update the interface of the Triangles window.